Make weight an Integer and include unit in label

+++ app/documents/renalware/transplants/recipient_operation_document.rb
@@ -8,7 +8,7 @@ module Renalware
         attribute :side, enums: %i(left right both)
         attribute :asystolic, enums: %i(yes no)
         attribute :age, Age
-        attribute :weight, Float
+        attribute :weight, Integer
 
         validates :weight, numericality: { allow_blank: true }
       end
